requirements.txt
Django==1.9.7
django-haystack==2.4.1
django-tastypie==0.13.3
elasticsearch==2.3.0
mysqlclient==1.3.7
python-dateutil==2.5.3
python-mimeparse==1.5.2
six==1.10.0
urllib3==1.16
